requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
pyclean==3.0.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
